Fix buffer overflow due to integer promotion rules
Integer promotion rules convert both sides of the comparison to float,
meaning the loss of precision on the float is not caught. This uncaught
imprecision can then result in width vs. size.width and height vs.
size.height being off by one for e.g. 2^24 + 1.

(cherry picked from commit 67b96e6a91c091445244abfa54e4f0572e26292f)

+++ src/OFBMPImageFormatHandler.m
@@ -49,7 +49,7 @@ byteSwapLine(void *line, size_t length, size_t byteSwapSize)
 	char magic[2];
 	uint32_t dataStart, headerSize, compressionMethod;
 	int32_t tmp32, horizPixelPerMeter, vertPixelPerMeter;
-	size_t width, height, lineLength, linePadding = 0;
+	uint32_t width, height, lineLength, linePadding = 0;
 	bool flipped = false;
 	OFSize size;
 	uint16_t bitsPerPixel;
@@ -96,7 +96,7 @@ byteSwapLine(void *line, size_t length, size_t byteSwapSize)
 	size.width = width;
 	size.height = height;
 
-	if (size.width != width || size.height != height)
+	if ((uint32_t)size.width != width || (uint32_t)size.height != height)
 		@throw [OFOutOfRangeException exception];
 
 	/* Number of color planes */
@@ -105,7 +105,7 @@ byteSwapLine(void *line, size_t length, size_t byteSwapSize)
 
 	bitsPerPixel = [stream readLittleEndianInt16];
 
-	if (SIZE_MAX / width < bitsPerPixel / CHAR_BIT)
+	if (UINT32_MAX / width < bitsPerPixel / CHAR_BIT)
 		@throw [OFOutOfRangeException exception];
 
 	lineLength = width * (bitsPerPixel / CHAR_BIT);
